+++ Logic/Models/MovieModel.cs
@@ -16,8 +16,8 @@ namespace Logic.Models
         public int MediaId { get; set; } = -1;
         public bool IsFavourite { get; set; } = false;
         public string ImagePath { get; set; }
-        public double AverageRating { get; set; }
-        public double PersonalRating { get; set; }
+        public int AverageRating { get; set; }
+        public int PersonalRating { get; set; }
 
         public MovieModel(int movieId, string title, string description, DateTime releaseDate, int mediaId)
         {
